[API Python] taking into account some remarks from reviewers
