+++ CCPAPI/cmake/FindPYQTSIP.cmake
@@ -16,7 +16,6 @@
 #                                                                        #
 ##########################################################################
 
-
 # - Find PYQT and SIP
 # Sets the following variables:
 #   PYQT_PYUIC_EXECUTABLE  - path to pyuic5 script           : pyuic5
@@ -35,9 +34,12 @@ endif()
 find_program(PYQT_PYUIC_EXECUTABLE NAMES pyuic5 pyuic5.bat DOC "full path of pyuic5 script")
 find_program(PYQT_PYRCC_EXECUTABLE NAMES pyrcc5 pyrcc5.bat DOC "full path of pyrcc5 script")
 
-execute_process(COMMAND ${PYTHON_EXECUTABLE} -c "import sys;
-hints=' '.join(sys.path);
-sys.stdout.write(hints)"
+execute_process(COMMAND ${PYTHON_EXECUTABLE} -c
+"
+import sys
+hints=' '.join(sys.path)
+sys.stdout.write(hints)
+"
 OUTPUT_VARIABLE _tmpPythonPath)
 
 separate_arguments(_tmpPythonPath)
@@ -54,16 +56,19 @@ find_path(PYQT_SIPS QtGui/QtGuimod.sip HINTS "/usr/share/sip/PyQt5" DOC "directo
 message(STATUS "PYQT_SIPS: " ${PYQT_SIPS})
 
 if(SIP_INCLUDE_DIR)
-  get_filename_component(SIP_PYTHONPATH "${SIP_INCLUDE_DIR}" PATH)
-  get_filename_component(SIP_PYTHONPATH "${SIP_PYTHONPATH}" PATH)
-  set(SIP_PYTHONPATH "${SIP_PYTHONPATH}/lib/python${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}/site-packages")
-  message(STATUS "SIP_PYTHONPATH: " ${SIP_PYTHONPATH})
+    get_filename_component(SIP_PYTHONPATH "${SIP_INCLUDE_DIR}" PATH)
+    get_filename_component(SIP_PYTHONPATH "${SIP_PYTHONPATH}" PATH)
+    set(SIP_PYTHONPATH "${SIP_PYTHONPATH}/lib/python${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}/site-packages")
+    message(STATUS "SIP_PYTHONPATH: " ${SIP_PYTHONPATH})
 endif()
 
-execute_process(COMMAND ${PYTHON_EXECUTABLE} -c "import sys; 
-sys.path[:0] = '${PYQT_PYTHONPATH}'.split(';');
-from PyQt5.QtCore import PYQT_CONFIGURATION;
-sys.stdout.write(PYQT_CONFIGURATION['sip_flags'])"
+execute_process(COMMAND ${PYTHON_EXECUTABLE} -c
+"
+import sys
+sys.path[:0] = '${PYQT_PYTHONPATH}'.split(';')
+from PyQt5.QtCore import PYQT_CONFIGURATION
+sys.stdout.write(PYQT_CONFIGURATION['sip_flags'])
+"
 OUTPUT_VARIABLE PYQT_SIPFLAGS)
 
 separate_arguments(PYQT_SIPFLAGS)
@@ -73,5 +78,3 @@ message(STATUS "PYQT_SIPFLAGS: " "${PYQT_SIPFLAGS}")
 
 include(FindPackageHandleStandardArgs)
 find_package_handle_standard_args(PYQTSIP REQUIRED_VARS PYQT_PYUIC_EXECUTABLE SIP_INCLUDE_DIR SIP_EXECUTABLE PYQT_SIPFLAGS)
-
-
