+++ CCPAPI/sip/ccPointCloud.sip
@@ -179,4 +179,3 @@ public:
     %End
     
 };
-
